Add whole project
The TTSBulkImporter works, but parameters/configuration must be provided
directly through code in `TTSBulkImporter/Program.cs`. There is room for
refactoring and modularity so that the app can be used by anyone for any
Tabletop Simulator mod. For now, much is hardcoded directly for the
Heroes System Tactical Scale mod.